refactor: modularize lib.rs by splitting into smaller focused modules
- Split 1227-line lib.rs into smaller, responsibility-based modules
- Created new module structure:
  - cli.rs: CLI argument parsing and command dispatch
  - commands/: Individual command implementations (build, check, clean)
  - doctor.rs: Environment diagnostic functionality
  - python.rs: PyO3 Python bindings
  - utils.rs: Common utility functions
  - verification/: Import verification logic (package_structure, script_generator, import_test)
- Reduced lib.rs from 1227 lines to 31 lines with clean module declarations
- All existing functionality preserved, tests pass (58 unit + 15 E2E tests)
- Improved code organization and maintainability
