+++ src/commands/mod.rs
@@ -1,0 +1,7 @@
+pub mod build;
+pub mod check;
+pub mod clean;
+
+pub use build::build;
+pub use check::check;
+pub use clean::clean;
